network/task-1/requirements.txt
pycryptodome==3.10.1
scapy
